requirements.txt
a2a-sdk==0.3.23
ag-ui-protocol==0.1.18
agent-framework==1.2.0
agent-framework-a2a==1.0.0b260424
agent-framework-ag-ui==1.0.0b260424
agent-framework-anthropic==1.0.0b260424
agent-framework-azure-ai-search==1.0.0b260424
agent-framework-azure-cosmos==1.0.0b260424
agent-framework-azurefunctions==1.0.0b260424
agent-framework-bedrock==1.0.0b260424
agent-framework-chatkit==1.0.0b260424
agent-framework-claude==1.0.0b260424
agent-framework-copilotstudio==1.0.0b260424
agent-framework-core==1.2.0
agent-framework-declarative==1.0.0b260424
agent-framework-devui==1.0.0b260424
agent-framework-durabletask==1.0.0b260424
agent-framework-foundry==1.2.0
agent-framework-foundry-local==1.0.0b260424
agent-framework-github-copilot==1.0.0b260424
agent-framework-lab==1.0.0b251024
agent-framework-mem0==1.0.0b260424
agent-framework-ollama==1.0.0b260424
agent-framework-openai==1.2.0
agent-framework-orchestrations==1.0.0b260424
agent-framework-purview==1.0.0b260424
agent-framework-redis==1.0.0b260424
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.80.0
anyio==4.13.0
asyncio==4.0.0
attrs==26.1.0
azure-ai-inference==1.0.0b9
azure-ai-projects==2.1.0
azure-common==1.1.28
azure-core==1.39.0
azure-cosmos==4.15.0
azure-functions==1.24.0
azure-functions-durable==1.5.0
azure-identity==1.25.3
azure-search-documents==11.7.0b2
azure-storage-blob==12.28.0
backoff==2.2.1
boto3==1.42.96
botocore==1.42.96
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
claude-agent-sdk==0.1.48
click==8.3.3
clr_loader==0.2.10
cryptography==47.0.0
distro==1.9.0
docstring_parser==0.18.0
durabletask==1.4.0
durabletask.azuremanaged==1.4.0
fastapi==0.133.0
foundry-local-sdk==0.5.1
frozenlist==1.8.0
furl==2.1.4
github-copilot-sdk==0.2.1
google-api-core==2.30.3
google-auth==2.49.2
googleapis-common-protos==1.74.0
griffelib==2.0.2
grpcio==1.80.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.13
importlib_metadata==8.7.1
isodate==0.7.2
Jinja2==3.1.6
jiter==0.14.0
jmespath==1.1.0
jsonpath-ng==1.8.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
mcp==1.27.0
mem0ai==1.0.11
microsoft-agents-activity==0.3.1
microsoft-agents-copilotstudio-client==0.3.1
microsoft-agents-hosting-core==0.3.1
ml_dtypes==0.5.4
msal==1.36.0
msal-extensions==1.3.1
multidict==6.7.1
numpy==2.4.4
ollama==0.5.3
openai==2.32.0
openai-agents==0.14.6
openai-chatkit==1.6.3
opentelemetry-api==1.41.1
opentelemetry-sdk==1.41.1
opentelemetry-semantic-conventions==0.62b1
orderedmultidict==1.0.2
packaging==26.2
portalocker==3.2.0
posthog==7.13.1
powerfx==0.0.34
propcache==0.4.1
proto-plus==1.27.2
protobuf==6.33.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.3
pydantic-settings==2.14.0
pydantic_core==2.46.3
PyJWT==2.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.26
python-ulid==3.1.0
pythonnet==3.0.5
pytz==2026.1.post1
PyYAML==6.0.3
qdrant-client==1.17.1
redis==7.1.1
redisvl==0.15.0
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
s3transfer==0.16.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.49
sse-starlette==3.3.4
starlette==1.0.0
tenacity==9.1.4
tqdm==4.67.3
types-requests==2.33.0.20260408
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
Werkzeug==3.1.8
yarl==1.23.0
zipp==3.23.1
